requirements.txt
python-telegram-bot==13.0
numpy>=1.16.5
opencv-python>=4.4.0.44